Move AppPage to separated file
