+++ src/app/components/Sidebar.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect } from 'react';
 import { MIN_MENU_TITLE_WIDTH } from '../layout-constants';
 import MyLibrary from './MyLibrary';
 import SidebarMenu from './SidebarMenu';
-import { AppPage } from '../types';
+import { AppPage } from '../types/app';
 import useMouseResize from '../customHook/useMouseResize';
 import { usePathname } from 'next/navigation';
