+++ src/app/types/app.ts
@@ -1,0 +1,9 @@
+export enum AppPage {
+  HOME = '/',
+  SEARCH = '/search',
+  LOGIN = '/login',
+  PLAYLIST = '/playlist',
+  ALBUM = '/album',
+  ARTIST = '/artist',
+  TRACK = '/track',
+}
